add express route to fetch one item
