+++ public/react/components/Item.js
@@ -3,8 +3,10 @@ import React from 'react';
 export const Item = (props) => {
 
   return <>
-    <h3>{props.item.name}</h3>
-    <img src={props.item.image} alt={props.item.name} />
+    <div id="itemslist">
+    <h3>{props.items.title}</h3>
+    <img src={props.items.image} alt={props.items.title} />
+    </div>
   </>
 }
