+++ public/style.css
@@ -12,3 +12,8 @@ body {
   border-radius: 1rem;
   height: 300px;
 }
+
+#itemslist{
+  background-color: aliceblue;
+  border: 10px solid red;
+}
