+++ server/routes/index.js
@@ -2,6 +2,6 @@ const express = require("express");
 const router = express.Router();
 
 // different model routers
-router.use('/sauces', require('./sauces'));
+router.use('/items', require('./items'));
 
 module.exports = router;
