--- server/routes/sauces.js
@@ -1,15 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const { Sauce } = require("../models");
-
-// GET /sauce
-router.get("/", async (req, res, next) => {
-  try {
-    const sauces = await Sauce.findAll();
-    res.send(sauces);
-  } catch (error) {
-    next(error);
-  }
-});
-
-module.exports = router;
